Parse metadata of remote images (except exif)

+++ app/models/plotline/image.rb
@@ -4,27 +4,34 @@ module Plotline
     after_destroy :remove_image_file
 
     def filename
-      File.join('./public', image)
+      remote_file? ? image : File.join('./public', image)
     end
 
     def remove_image_file
-      File.delete(filename)
+      File.delete(filename) unless remote_file?
     end
 
     private
 
     def set_metadata
+      img = FastImage.new(filename)
+
+      self.width, self.height = img.size
+      self.ratio = self.width.to_f / self.height.to_f
+      self.content_type = img.type
+      self.file_size = img.content_length
+
+      return if remote_file?
+
       File.open(filename) do |file|
-        img = FastImage.new(file)
-
-        self.width, self.height = img.size
-        self.ratio = self.width.to_f / self.height.to_f
-
-        self.content_type = img.type
-        self.file_size    = file.size
+        self.file_size = file.size # content_length doesn't always work
       end
 
       self.exif = Exiftool.new(filename).to_hash
+    end
+
+    def remote_file?
+      image.start_with?('http')
     end
   end
 end
